requirements.txt
python-telegram-bot==21.4
sqlalchemy==2.0.25
pandas==2.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
matplotlib>=3.8
Pillow>=10.0
requests==2.32.3
loguru==0.7.2

Flask
gunicorn
Flask-Cors

yfinance
pycoingecko
yahoo_fin==0.8.9.1
beautifulsoup4==4.12.3
lxml==5.3.0
html5lib==1.1
python-dateutil==2.9.0.post0
numpy

aiohttp
certifi

telethon==1.34.0
pytz==2024.1
